Move `GroceryStub` to dedicated file
